Commit working code
Initial commit of working code which successfully connects an ESP32 to a Micromelon Rover wireless via Bluetooth Low Energy. The code implements functionality to control the movement of the rover via four pushbuttons connected to the GPIO pins of the ESP32 to make it move forward and backwards as well as pivot on the spot in a clockwise and counter-clockwise fashion.